requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
icalendar==5.0.11
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonrpcclient==4.0.3
jsonschema==4.19.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
MarkupSafe==2.1.5
marshmallow==3.20.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.11.1
packaging==23.2
pandas==2.0.3
pkgutil_resolve_name==1.3.10
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
tabulate==0.9.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
